+++ cognite/neat/rules/transformers/_mapping.py
@@ -2,9 +2,11 @@ from abc import ABC
 from collections import defaultdict
 
 from cognite.neat.rules._shared import JustRules, OutRules
-from cognite.neat.rules.models import DMSRules
+from cognite.neat.rules.models import DMSRules, InformationRules
+from cognite.neat.rules.models._base_rules import ClassRef
 from cognite.neat.rules.models.dms import DMSProperty
 from cognite.neat.rules.models.entities import ReferenceEntity
+from cognite.neat.rules.models.mapping import RuleMapping
 
 from ._base import RulesTransformer
 
@@ -95,3 +97,37 @@ class MapOneToOne(MapOntoTransformers):
                 prop.reference = ReferenceEntity.from_entity(ref_prop.view, ref_prop.view_property)
 
         return JustRules(solution)
+
+
+class RuleMapper(RulesTransformer[InformationRules, InformationRules]):
+    """Maps properties and classes using the given mapping.
+
+    **Note**: This transformer mutates the input rules.
+
+    Args:
+        mapping: The mapping to use.
+
+    """
+
+    def __init__(self, mapping: RuleMapping) -> None:
+        self.mapping = mapping
+
+    def transform(self, rules: InformationRules | OutRules[InformationRules]) -> JustRules[InformationRules]:
+        input_rules = self._to_rules(rules)
+
+        destination_by_source = self.mapping.properties.as_destination_by_source()
+        destination_cls_by_source = self.mapping.classes.as_destination_by_source()
+        for prop in input_rules.properties:
+            if destination_prop := destination_by_source.get(prop.as_reference()):
+                prop.class_ = destination_prop.class_
+                prop.property_ = destination_prop.property_
+            elif destination_cls := destination_cls_by_source.get(ClassRef(Class=prop.class_)):
+                # If the property is not in the mapping, but the class is,
+                # then we should map the class to the destination
+                prop.class_ = destination_cls.class_
+
+        for cls_ in input_rules.classes:
+            if destination_cls := destination_cls_by_source.get(cls_.as_reference()):
+                cls_.class_ = destination_cls.class_
+
+        return JustRules(input_rules)