+++ docs/CHANGELOG.md
@@ -31,6 +31,7 @@ Changes are grouped as follows:
 ### Added
 - Rules exporter that produces a spreadsheet template for instance creation based on definition of classes in the rules
 - Rules transformer which converts information rules entities to be DMS compliant
+- Rules transformer `RuleMapping` that maps rules from one data model to another
 
 ## [0.92.3] - 17-09-24
 ### Fixed